Carve out class methods of Haml::Compiler
to Haml::AttributeBuilder.

They are different responsibilities.
