+++ lib/haml/buffer.rb
@@ -194,7 +194,7 @@ module Haml
         self.class.merge_attrs(attributes, Hash[old.map {|k, v| [k.to_s, v]}])
       end
       self.class.merge_attrs(attributes, parse_object_ref(obj_ref)) if obj_ref
-      Compiler.build_attributes(
+      AttributeBuilder.build_attributes(
         html?, @options[:attr_wrapper], @options[:escape_attrs], @options[:hyphenate_data_attrs], attributes)
     end
 
@@ -223,14 +223,14 @@ module Haml
     # @param from [{String => #to_s}] The attribute hash to merge from
     # @return [{String => String}] `to`, after being merged
     def self.merge_attrs(to, from)
-      from[ID_KEY] = Compiler.filter_and_join(from[ID_KEY], '_') if from[ID_KEY]
+      from[ID_KEY] = AttributeBuilder.filter_and_join(from[ID_KEY], '_') if from[ID_KEY]
       if to[ID_KEY] && from[ID_KEY]
         to[ID_KEY] << "_#{from.delete(ID_KEY)}"
       elsif to[ID_KEY] || from[ID_KEY]
         from[ID_KEY] ||= to[ID_KEY]
       end
 
-      from[CLASS_KEY] = Compiler.filter_and_join(from[CLASS_KEY], ' ') if from[CLASS_KEY]
+      from[CLASS_KEY] = AttributeBuilder.filter_and_join(from[CLASS_KEY], ' ') if from[CLASS_KEY]
       if to[CLASS_KEY] && from[CLASS_KEY]
         # Make sure we don't duplicate class names
         from[CLASS_KEY] = (from[CLASS_KEY].to_s.split(' ') | to[CLASS_KEY].split(' ')).sort.join(' ')
